% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StudySpecific.R
\name{execute}
\alias{execute}
\title{Execute OHDSI Study phenotypeDataExtraction}
\usage{
execute(connectionDetails, cdm, results, workFolder, ncnt = 450, mcnt = 450,
  oracleTempSchema = results, ...)
}
\arguments{
\item{cdm}{Schema name where your patient-level data in OMOP CDM format resides}

\item{results}{Schema where you'd like the results tables to be created (requires user to have create/write access)}
}
\value{
Study results are in folder
}
\details{
This function executes OHDSI Study (FILL IN NAME).
This is a study of (GIVE SOME DETAILS).
Detailed information and protocol are available on the OHDSI Wiki.
}

