% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phenotypeDataExtraction-package.R
\docType{package}
\name{phenotypeDataExtraction}
\alias{phenotypeDataExtraction}
\alias{phenotypeDataExtraction-package}
\title{phenotypeDataExtraction.}
\description{
phenotypeDataExtraction.
}

